improve Embed image button UX
